requirements.txt
pyvmomi>=7.0
PyYAML>=6.0
matplotlib>=3.8.0
